Added memory range constants.
Some cleanup in Lexer. The motivation for the memory constant file is so
that we can implement something that lets us dereference from the data
segment. If we could do this, then we can write

.data
ten: .word 10

.text
lui $t0 ten

and the value in $t0 will be 0xA (the value at the address labelled with
'ten'). This means that at least the resolver function needs to be able
to tell if the label is in the text or data section, and we need to be
able to query the data section info to get the value at some address.